py/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
marshmallow==3.20.1
python-dotenv==1.0.0
pytz==2023.3
pytest==7.4.2
Werkzeug==2.3.7
SQLAlchemy==2.0.20
PyJWT==2.8.0
Flask-Cors==4.0.0
PyMySQL==1.1.0
